requirements.txt
# ── Core Framework ─────────────────────────────────────────
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-migrate>=4.0.0
flask-wtf>=1.2.0
flask-login>=0.6.2
werkzeug>=3.0.0
python-dotenv>=1.0.0

# ── Database ───────────────────────────────────────────────
psycopg2-binary>=2.9.0

# ── ML / Data Science ─────────────────────────────────────
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
lightgbm>=4.1.0

# ── AI / LLM ──────────────────────────────────────────────
google-genai>=1.0.0

# ── Security & Auth ───────────────────────────────────────
bcrypt>=4.1.2
cryptography>=42.0.0

# ── HTTP / OAuth ──────────────────────────────────────────
requests>=2.31.0

# ── Production Server ─────────────────────────────────────
gunicorn>=22.0.0
